refactor: 하나의 클래스에서 관리하도록 변경
